refactor(domain): enum 클래스들의 패키지 이동
domain 패키지 내부에 클래스들이 너무 많아 열거형 클래스들을 따로 enums 패키지 내부로 이동했습니다.
