+++ src/main/java/domain/Game.java
@@ -1,5 +1,6 @@
 package domain;
 
+import domain.enums.Result;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
