+++ src/main/java/domain/Participant.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.enums.Rank;
+import domain.enums.Result;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
