+++ src/main/java/domain/enums/Result.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.enums;
 
 public enum Result {
