+++ src/main/java/domain/enums/Suit.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.enums;
 
 public enum Suit {
     HEART("하트"),
